backend/requirements.txt
# Backend Requirements
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
pydantic>=2.5.2
python-multipart>=0.0.6
yt-dlp>=2023.11.16

# Transcription
faster-whisper>=0.9.0

# Translation
argostranslate>=1.9.1

# Export
# (no additional deps)

# Speaker Diarization (optional - requires HF_TOKEN)
# pyannote.audio>=3.1.0
# torch>=2.0.0

# LLM Content Repurposing
google-generativeai>=0.3.0

# TTS Dubbing
edge-tts>=6.1.9

# CORS
# (included in fastapi)
